Downgrade .NET to 4.8.1
.NET 7 is cool and all but I don't think there's much of a performance difference. Requiring manual install of .NET 7 to even launch isn't exactly user friendly. Since Windows includes net framework already, better to allow the app be standalone and just immediately open without anything extra.

Will upgrade back to .NET 7 when WPF supports trimming and .NET self-contained. (.NET 8?) Alternatively if WPF is ever replaced with a new UI (Avalonia) instead.

+++ BardMusicPlayer.Seer/Game.cs
@@ -225,9 +225,9 @@ public partial class Game : IDisposable, IEquatable<Game>
 
     public static bool operator !=(Game game, Game otherGame) => game is not null && !game.Equals(otherGame);
 #pragma warning disable CA1416
-    public IntPtr GetAffinity()
+    public long GetAffinity()
     {
-        return Process.ProcessorAffinity;
+        return (long)Process.ProcessorAffinity;
     }
 
     public void SetAffinity(long AffinityMask)
